# docs/swagger.yaml
definitions:
...
  datastore.Card:
    properties:
      approved:
        type: string
      company_id:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      linkToProd:
        type: string
      name:
        type: string
      price:
        type: string
      tags:
        items:
          type: string
        type: array
      views:
        type: integer
    type: object
  datastore.Company:
    properties:
      cards:
        items:
          $ref: '#/definitions/datastore.Card'
        type: array
      deletedAt:
        type: string
      id:
        type: string
      image:
        type: string
      isDel:
        type: integer
      linkToCompany:
        type: string
      mail:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  utility.JWTTokenPair:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  utility.Message:
    properties:
      text:
        type: string
    type: object
host: http://localhost:8081
info:
  contact: {}
  description: Nenavijy swagger
  license:
    name: BSD-3
  title: pay-with-crypto API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to company account.
      parameters:
      - description: Company login data
        in: body
        name: login_data
        required: true
        schema:
          properties:
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.JWTTokenPair'
        "400":
          description: Invalid company email
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Company already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create company.
      parameters:
      - description: Company data
        in: body
        name: company_data
        required: true
        schema:
          properties:
            linkToCompany:
              type: string
            mail:
              type: string
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/datastore.Company'
        "400":
          description: Invalid company email
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Company already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Auth
...
  /auth/token_update:
    post:
      consumes:
      - application/json
      description: Update tokens.
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.JWTTokenPair'
        "400":
          description: Can't update refresh token
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Token already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Auth
...
securityDefinitions:
  string:
    in: header
    name: accessToken
    type: apiKey
swagger: "2.0"